Support ergonomic foreign key declarations
