+++ docs/todo/todo.md
@@ -11,7 +11,7 @@
 - [x] Live-workspace verification for column renames. RESOLVED 2026-07-14 (manual notebook check against a real workspace): (1) Unity Catalog column tags and comments travel with `ALTER TABLE ... RENAME COLUMN` (tag observed under the new name, comment intact), so the differ's projected comparison is correct and a rename converges in a single sync; (2) a dependent CHECK constraint blocks the rename with `DELTA_CONSTRAINT_DEPENDENT_COLUMN_CHANGE` — a clean execution error, per the no-preflight policy; (3) the original suppression question is obsolete (the plan states PK/FK drops explicitly since the explicit-drops refactor), and the underlying hazard is confirmed real: a direct `RENAME COLUMN` on a PK column with a live inbound FK succeeds and silently drops both the PK and the other table's FK, while the engine's bare `DROP PRIMARY KEY IF EXISTS` fails closed first (`DependentConstraintsExistException`: "Cannot delete constraint ... with non empty child constraints"). Also verified: the engine's exact three-statement PK-rename sequence (drop key, rename, re-add key) executes end-to-end, and a partition-column rename leaves data and partition layout intact. Residual unknown: governed tags (docs say they block `DROP COLUMN`; behaviour under a rename is untested — the verified travel was a regular tag).
 - [ ] Cross-catalog inbound foreign keys are invisible to `PrimaryKeyReferencedByForeignKeys` (information_schema is per-catalog), so a cross-catalog referenced PK drop still fails at execution rather than validation. Decide whether that is acceptable long-term or worth probing the referencing catalogs (cost: catalog reads fan out). The outbound direction is rejected at declaration for the same reason (`ForeignKey` requires the referenced table to share the owner's catalog); lifting either restriction means resolving the parent side from the parent's catalog or the metastore-wide `system.information_schema`.
 - [ ] Watch the inbound-FK read cost: observing referencing foreign keys adds one information_schema query per table read (PR #152). No latency evidence yet — do not optimize without it; if large syncs hurt, batch the metadata reads per catalog rather than per table.
-- [ ] Release notes for the next release must call out the stricter property values: previously-accepted declarations like `"True"`, `"1_000"`, or compound intervals (`interval 1 hour 30 minutes`) now fail at declaration time with messages naming the expected format. Also call out two breaking declaration changes from the FK-pairing redesign: `primary_key` moved from a per-column `Column(primary_key=True)` flag to a `DeltaTable(primary_key=[...])` parameter, and `ForeignKey(columns={local: referenced})` replaces the positional `local_columns=[...]` list. Both old forms fail loudly at construction (unexpected keyword argument, or a missing/extra-column `ValueError`) rather than silently changing meaning. Also call out: `delta.enableTypeWidening` is now a managed key — per the registry admission policy this is breaking, since tables with the key set but undeclared start failing validation (`PropertyMustBeDeclared`) on upgrade.
+- [ ] Release notes for the next release must call out the stricter property values: previously-accepted declarations like `"True"`, `"1_000"`, or compound intervals (`interval 1 hour 30 minutes`) now fail at declaration time with messages naming the expected format. Also call out two breaking declaration changes from the FK-pairing redesign: `primary_key` moved from a per-column `Column(primary_key=True)` flag to a `DeltaTable(primary_key=[...])` parameter, and `ForeignKey(columns=...)` replaces the positional `local_columns=[...]` parameter. The new `columns` form accepts a string for a single-column parent key, a sequence for a same-name key, or an explicit `{local: referenced}` mapping for an ambiguous composite key; the old `local_columns` keyword fails loudly as an unexpected argument rather than silently changing meaning. Also call out: `delta.enableTypeWidening` is now a managed key — per the registry admission policy this is breaking, since tables with the key set but undeclared start failing validation (`PropertyMustBeDeclared`) on upgrade.
 - [ ] add unique columns: ALTER TABLE U ADD CONSTRAINT u_uq_email UNIQUE(email);
 - [ ] Make the application layer backend-agnostic, not just backend-library-free. Today the `backend-imports-stay-in-adapters` contract only forbids importing the `delta`/`pyspark` _packages_; the application layer still hardcodes Delta-family _policy_ — the safe-change rules (`PartitioningChangeNotSupported`, `ColumnDataTypeChangeNotSupported`, `NonNullableColumnAdd`, `ColumnMappingRequiredForDrop`) exist because Databricks/Delta rejects those operations, and `application/properties.py` holds the Delta property catalogue (`DELTA_PROPERTY_REGISTRY`, `COLUMN_MAPPING_MODE_KEY`, `delta.*` key strings) consumed by `validation.py`, `api/properties.py`, and the reader. The layering is sound for now (application is the only layer all three consumers can reach without violating the `adapters|api → application → domain` order), but it means "backend-agnostic" is only half true. The real refactor: `application` holds only the _mechanism_ (how rules run, the `Rule` protocol, how `validate_diff` composes them), and the backend _specifics_ (which rules, which property keys and their restrictions) are injected from the composition root or an adapter-provided port. This must move ALL the backend-flavoured rules and the property catalogue together — moving only the property pieces just relocates a third of the problem. Larger change, own PR. See the property-ownership spec addendum for how the registry ended up in application.
 - [ ] Think of whether to make DeltaTable automatically make columns unique if they are PK cols
