+++ src/delta_engine/api/delta_table.py
@@ -303,11 +303,13 @@ class ForeignKey:
     """
     Public declaration of a foreign key relationship.
 
-    ``columns`` maps each local column to the referenced-table primary key
-    column it references — the mapping is the pairing, stated outright; its
-    insertion order is irrelevant. The values must cover the referenced
-    table's primary key exactly. The physical constraint name is generated by
-    the engine and is not part of this declaration.
+    ``columns`` accepts one local column name for a single-column parent
+    key, a sequence of local names for a same-name key, or an explicit
+    ``{local: referenced}`` mapping. Sequence and mapping orders are irrelevant;
+    ambiguous composite keys require the explicit form. Identifiers are
+    normalized to lowercase when the declaration is constructed. The physical
+    constraint name is generated by the engine and is not part of this
+    declaration.
 
     ``references`` is another :class:`DeltaTable`, or the :data:`Self` sentinel
     for a self-referential key. See the architecture explanation doc for why
@@ -317,17 +319,35 @@ class ForeignKey:
     observed afterwards.
     """
 
-    columns: Mapping[str, str]
+    columns: str | Sequence[str] | Mapping[str, str]
     references: DeltaTable | _SelfReference
 
     def __post_init__(self) -> None:
-        if not isinstance(self.columns, Mapping):
-            raise TypeError(
-                "foreign key columns must be a mapping of"
-                " {local column: referenced column};"
-                f" got {type(self.columns).__name__}"
-            )
-        object.__setattr__(self, "columns", MappingProxyType(dict(self.columns)))
+        # Freeze and normalize user input once; resolve it when the owning
+        # DeltaTable is constructed.
+        frozen: tuple[str, ...] | Mapping[str, str]
+        match self.columns:
+            case str():
+                frozen = (self.columns.lower(),)
+            case Mapping():
+                frozen = MappingProxyType(
+                    {
+                        local.lower(): referenced.lower()
+                        for local, referenced in self.columns.items()
+                    }
+                )
+            case Sequence():
+                if not all(isinstance(column, str) for column in self.columns):
+                    raise TypeError("foreign key columns must be strings")
+                frozen = tuple(column.lower() for column in self.columns)
+            case _:
+                raise TypeError(
+                    "foreign key columns must be a column name,"
+                    " a sequence of same-name columns, or"
+                    " a {local: referenced} mapping"
+                )
+
+        object.__setattr__(self, "columns", frozen)
 
     def _to_constraint(
         self,
@@ -339,8 +359,8 @@ class ForeignKey:
         Lower this declaration into a domain constraint.
 
         Applies the lowering rules in order: the referenced table must live in
-        the owner's catalog, must declare a primary key, the mapping's values
-        must equal that key exactly, and each local column's data type must
+        the owner's catalog, must declare a primary key, the resolved parent
+        columns must equal that key exactly, and each local column's data type must
         match its referenced column's. Local column existence is not checked
         here — the ``DesiredTable`` built right after enforces it.
         """
@@ -360,8 +380,9 @@ class ForeignKey:
                 f"foreign key references {referenced.table}, which declares no primary key"
             )
 
-        local_columns = tuple(self.columns.keys())
-        referenced_columns = tuple(self.columns.values())
+        pairs = self._resolve_column_pairs(referenced)
+        local_columns = tuple(local for local, _ in pairs)
+        referenced_columns = tuple(parent for _, parent in pairs)
 
         declared = set(referenced_columns)
         key = set(referenced.key_columns)
@@ -380,7 +401,7 @@ class ForeignKey:
             )
 
         local_types = {column.name: column.data_type for column in owner_columns}
-        for local_name, referenced_name in self.columns.items():
+        for local_name, referenced_name in pairs:
             local_type = local_types.get(local_name)
             if local_type is None:
                 continue  # local column existence is enforced when the DesiredTable is built
@@ -427,6 +448,31 @@ class ForeignKey:
                 raise TypeError(
                     f"foreign key references must be a DeltaTable or Self; got {self.references!r}"
                 )
+
+    def _resolve_column_pairs(
+        self,
+        referenced: _ReferencedSide,
+    ) -> tuple[tuple[str, str], ...]:
+        """Resolve the declaration into explicit local-to-parent column pairs."""
+        if isinstance(self.columns, Mapping):
+            return tuple(self.columns.items())
+
+        local_columns = tuple(self.columns)
+        parent_columns = referenced.key_columns
+
+        # a single-column parent has only one possible pairing
+        if len(local_columns) == 1 and len(parent_columns) == 1:
+            return ((local_columns[0], parent_columns[0]),)
+
+        if set(local_columns) == set(parent_columns):
+            return tuple((column, column) for column in local_columns)
+
+        raise ValueError(
+            f"cannot infer foreign key column pairing for {referenced.table};"
+            f" local columns are ({', '.join(local_columns)}) and the referenced"
+            f" primary key is ({', '.join(parent_columns)}). Provide an explicit"
+            " mapping of {local column: referenced column}."
+        )
 
 
 class DeltaTable:
